+++ search/search_tests/keyword_matcher_test.cpp
@@ -260,7 +260,8 @@ string GetManyTokens(string tokenPrefix, int tokenCount, bool countForward = tru
 UNIT_TEST(KeywordMatcher_QueryTooLong)
 {
   static_assert(kMaxNumTokens >= 2, "");
-  for (int queryLength = kMaxNumTokens - 2; queryLength <= kMaxNumTokens + 2; ++queryLength)
+  for (int queryLength = kMaxNumTokens - 2;
+       queryLength <= static_cast<int>(kMaxNumTokens + 2); ++queryLength)
   {
     string const query = GetManyTokens("Q", queryLength);
     string const queryWithPrefix = query + " Prefix";
